fix(#1406): extend materialization-vs-wake invariant to legacy seeds and split admission module
Why:

- Legacy proposal seeds (pre-idempotency-key index) now participate in the same materialization-vs-wake-completion invariant: reconcile checks isDispatchComplete and re-dispatches an unwoken legacy row exactly once instead of returning legacySeed=true and leaving it permanently asleep.

- Extracted queue seed admission/dispatch state into packages/api/src/routes/proposal-seed-admission.ts, bringing proposal-approve-dispatch.ts back under the 350-line architecture contract.

- Added regression test for legacy queue-full seed repair: one message, one processNext, terminal delivered state, empty queue after delivery.

- Verified: pnpm --filter @cat-cafe/api build + 34 targeted proposal tests all pass.
