+++ packages/api/src/routes/proposal-approve-dispatch.ts
@@ -1,7 +1,6 @@
-import type { CatId, MessageContent, ReportingMode } from '@cat-cafe/shared';
-import type { InvocationQueue, QueueEntry } from '../domains/cats/services/agents/invocation/InvocationQueue.js';
+import type { CatId, ReportingMode } from '@cat-cafe/shared';
+import type { InvocationQueue } from '../domains/cats/services/agents/invocation/InvocationQueue.js';
 import type { OwnerAuthProvenance } from '../domains/cats/services/agents/invocation/owner-auth-provenance.js';
-import { createInitialQueuedMessageCustody } from '../domains/cats/services/agents/invocation/QueuedMessageCustodyCoordinator.js';
 import type { QueueProcessor } from '../domains/cats/services/agents/invocation/QueueProcessor.js';
 import { parseIntent } from '../domains/cats/services/context/IntentParser.js';
 import type { AgentRouter } from '../domains/cats/services/index.js';
@@ -10,7 +9,12 @@ import type { IThreadStore } from '../domains/cats/services/stores/ports/ThreadS
 import type { SocketManager } from '../infrastructure/websocket/index.js';
 import { primaryMentionHandleForCatId } from '../utils/cat-mention-handle.js';
 import { enrichWithParentThreadHeader } from './proposal-enrich-header.js';
-import { admitThreadParticipants } from './thread-participant-admission.js';
+import {
+  buildSourceEnvelopeContent,
+  executeQueuedDispatch,
+  resolveSourceContentBlocks,
+  type SourceEnvelope,
+} from './proposal-seed-admission.js';
 
 export { enrichWithParentThreadHeader } from './proposal-enrich-header.js';
 
@@ -88,52 +92,17 @@ export interface AppendApprovedInitialMessageInput extends ProposalInitialMessag
   messageStore: IMessageStore;
   threadStore: Pick<IThreadStore, 'addParticipants' | 'get'>;
   socketManager: Pick<SocketManager, 'emitToUser'>;
+  /**
+   * Optional previously-materialized seed (legacy or queue-full) to reuse instead
+   * of appending a new message. Used by reconcile to repair an existing row
+   * under the materialization-vs-wake-completion invariant.
+   */
+  existingSeed?: StoredMessage;
 }
 
 export interface AppendApprovedInitialMessageResult {
   messageId: string;
   warning?: string;
-}
-
-/**
- * Lossless source envelope delivered to a child thread when the proposal has no
- * explicit initialMessage. It carries the exact proposal fields (title, reason,
- * sourceMessageId) so the child can verify the original input instead of parsing
- * it out of the thread title.
- */
-export interface SourceEnvelope {
-  title: string;
-  reason: string;
-  /** Exact trigger message in the parent thread, if known. */
-  sourceMessageId?: string | null;
-}
-
-function buildSourceEnvelopeContent(envelope: SourceEnvelope): string {
-  const lines = [`**来源**: ${envelope.title}`, '', envelope.reason];
-  if (envelope.sourceMessageId) {
-    lines.push('', `**源消息**: \`${envelope.sourceMessageId}\``);
-  }
-  return lines.join('\n');
-}
-
-async function resolveSourceContentBlocks(
-  sourceEnvelope: SourceEnvelope,
-  messageStore: IMessageStore,
-): Promise<readonly MessageContent[] | undefined> {
-  if (!sourceEnvelope.sourceMessageId) {
-    return undefined;
-  }
-  try {
-    const sourceMessage = await messageStore.getById(sourceEnvelope.sourceMessageId);
-    if (sourceMessage?.contentBlocks && sourceMessage.contentBlocks.length > 0) {
-      return sourceMessage.contentBlocks;
-    }
-  } catch {
-    // P2: source-block lookup is best-effort. The approve route has already
-    // finalized the proposal; a transient storage failure here must not drop
-    // the dispatch. The text envelope is sufficient to wake the assigned cat.
-  }
-  return undefined;
 }
 
 interface DispatchPlan {
@@ -200,6 +169,7 @@ export async function appendApprovedInitialMessage({
   router,
   invocationQueue,
   queueProcessor,
+  existingSeed,
 }: AppendApprovedInitialMessageInput): Promise<AppendApprovedInitialMessageResult> {
   // F128 source envelope: seed content uses the envelope only when there is no
   // explicit user-typed initialMessage. Routing/intent must never see the
@@ -238,6 +208,15 @@ export async function appendApprovedInitialMessage({
   // Phase AA (AC-AA6): resolve source cat handle for routing credentials
   const sourceCatHandle = sourceCatId ? (primaryMentionHandleForCatId(sourceCatId) ?? `@${sourceCatId}`) : null;
   if (!router || !invocationQueue || !queueProcessor) {
+    if (existingSeed) {
+      // We cannot wake a target without dispatch dependencies. Cancel the existing
+      // seed so the reconcile loop stops retrying a permanently unwakeable row.
+      await messageStore.markCanceled(existingSeed.id);
+      return {
+        messageId: existingSeed.id,
+        warning: 'initialMessage dispatch skipped: routing dependencies unavailable (existing seed canceled)',
+      };
+    }
     const enrichedFallback = enrichWithParentThreadHeader(
       seedContent,
       sourceThreadId,
@@ -298,6 +277,15 @@ export async function appendApprovedInitialMessage({
   );
 
   if (targetCats.length === 0) {
+    if (existingSeed) {
+      // No resolvable target for an existing seed: cancel it so the invariant
+      // stays terminal and retries do not loop.
+      await messageStore.markCanceled(existingSeed.id);
+      return {
+        messageId: existingSeed.id,
+        warning: 'initialMessage dispatch skipped: no target cats resolved (existing seed canceled)',
+      };
+    }
     const stored = await messageStore.append({
       userId,
       catId: sourceCatId ?? null, // AC-AA4
@@ -336,200 +324,6 @@ export async function appendApprovedInitialMessage({
     socketManager,
     invocationQueue,
     queueProcessor,
+    existingSeed,
   });
-}
-
-interface ExecuteQueuedDispatchInput {
-  proposalId: string;
-  userId: string;
-  ownerAuthProvenance: OwnerAuthProvenance;
-  threadId: string;
-  content: string;
-  targetCats: readonly CatId[];
-  intentName: string;
-  sourceCatId: CatId | null | undefined;
-  crossPostExtra: Record<string, unknown>;
-  sourceContentBlocks: readonly MessageContent[] | undefined;
-  messageStore: IMessageStore;
-  threadStore: Pick<IThreadStore, 'addParticipants' | 'get'>;
-  socketManager: Pick<SocketManager, 'emitToUser'>;
-  invocationQueue: ProposalInvocationQueue;
-  queueProcessor: ProposalQueueProcessor;
-}
-
-interface ExistingSeedAdmissionResult {
-  kind: 'complete' | 'admitted' | 'failed';
-  messageId: string;
-  warning?: string;
-}
-
-/**
- * #1406 B1 Part 2: a previously materialized seed (e.g. from a queue-full
- * fallback append) must be atomically admitted into the queue so it can be
- * driven to a terminal delivery state. Without this, the seed stays in the
- * ambiguous undefined-deliveryStatus state forever and every reconcile retries
- * a brand-new enqueue, risking duplicate invocations once the queue has space.
- */
-async function ensureExistingSeedAdmitted(
-  existingSeed: StoredMessage,
-  entry: QueueEntry,
-  messageStore: IMessageStore,
-): Promise<ExistingSeedAdmissionResult> {
-  if (existingSeed.deliveryStatus === 'delivered' || existingSeed.deliveryStatus === 'canceled') {
-    return { kind: 'complete', messageId: existingSeed.id };
-  }
-
-  if (existingSeed.deliveryStatus === 'queued' && existingSeed.queueCustody) {
-    return { kind: 'admitted', messageId: existingSeed.id };
-  }
-
-  if (existingSeed.deliveryStatus !== 'queued') {
-    const prepared = await messageStore.prepareQueueAdmission(existingSeed.id);
-    if (prepared.kind === 'conflict') {
-      // A concurrent transition may have finished the seed; refresh once.
-      const refreshed = await messageStore.getById(existingSeed.id);
-      if (refreshed?.deliveryStatus === 'delivered' || refreshed?.deliveryStatus === 'canceled') {
-        return { kind: 'complete', messageId: refreshed.id };
-      }
-      return {
-        kind: 'failed',
-        messageId: existingSeed.id,
-        warning: 'existing seed could not be admitted to queue',
-      };
-    }
-  }
-
-  if (!existingSeed.queueCustody) {
-    const custody = createInitialQueuedMessageCustody(entry);
-    const initialized = await messageStore.initializeQueueCustody(existingSeed.id, custody);
-    if (initialized.kind !== 'initialized' && initialized.kind !== 'existing') {
-      return {
-        kind: 'failed',
-        messageId: existingSeed.id,
-        warning: 'existing seed custody initialization failed',
-      };
-    }
-  }
-
-  return { kind: 'admitted', messageId: existingSeed.id };
-}
-
-async function executeQueuedDispatch({
-  proposalId,
-  userId,
-  ownerAuthProvenance,
-  threadId,
-  content,
-  targetCats,
-  intentName,
-  sourceCatId,
-  crossPostExtra,
-  sourceContentBlocks,
-  messageStore,
-  threadStore,
-  socketManager,
-  invocationQueue,
-  queueProcessor,
-}: ExecuteQueuedDispatchInput): Promise<AppendApprovedInitialMessageResult> {
-  const idempotencyKey = `proposal-initial:${proposalId}`;
-  const enqueueResult = invocationQueue.enqueue({
-    threadId,
-    userId,
-    ownerAuthProvenance,
-    idempotencyKey,
-    content,
-    source: 'user',
-    targetCats: targetCats as CatId[],
-    intent: intentName,
-  });
-
-  if (enqueueResult.outcome === 'full' || !enqueueResult.entry) {
-    // Queue-full: persist a materialized seed without queue custody. It is
-    // intentionally NOT dispatch-complete so reconcile can CAS-admit it later.
-    const stored = await messageStore.append({
-      userId,
-      catId: sourceCatId ?? null, // AC-AA4
-      content,
-      mentions: [...targetCats],
-      timestamp: Date.now(),
-      threadId,
-      idempotencyKey,
-      extra: crossPostExtra, // AC-AA5
-      contentBlocks: sourceContentBlocks,
-    });
-    return {
-      messageId: stored.id,
-      warning: 'initialMessage dispatch skipped: queue is full',
-    };
-  }
-
-  let storedMessageId = enqueueResult.entry.messageId ?? null;
-  if (!enqueueResult.deduped || !storedMessageId) {
-    // A prior queue-full or failed-persistence attempt may have already
-    // materialized the seed. Reuse it atomically instead of appending a second
-    // message.
-    const existingSeed = await messageStore.getByIdempotencyKey(userId, threadId, idempotencyKey);
-    if (existingSeed) {
-      const admission = await ensureExistingSeedAdmitted(existingSeed, enqueueResult.entry, messageStore);
-      if (admission.kind === 'complete') {
-        return { messageId: admission.messageId };
-      }
-      if (admission.kind === 'failed') {
-        invocationQueue.rollbackEnqueue(threadId, userId, enqueueResult.entry.id);
-        return { messageId: admission.messageId, warning: admission.warning };
-      }
-      storedMessageId = admission.messageId;
-    } else {
-      try {
-        const stored = await messageStore.append({
-          userId,
-          catId: sourceCatId ?? null, // AC-AA4
-          content,
-          mentions: [...targetCats],
-          timestamp: Date.now(),
-          threadId,
-          idempotencyKey,
-          deliveryStatus: 'queued',
-          queueCustody: createInitialQueuedMessageCustody(enqueueResult.entry),
-          extra: crossPostExtra, // AC-AA5
-          contentBlocks: sourceContentBlocks,
-        });
-        storedMessageId = stored.id;
-      } catch (err) {
-        invocationQueue.rollbackEnqueue(threadId, userId, enqueueResult.entry.id);
-        throw err;
-      }
-    }
-    invocationQueue.backfillMessageId(threadId, userId, enqueueResult.entry.id, storedMessageId);
-  }
-
-  // F128 owns the final dispatch plan: preferredCats ordering and explicit
-  // parallel intent can differ from raw-message mentions. Admit only those
-  // final targets, after Queue accepted a durable carrier and before the
-  // processor can start, so Sidebar C2 and C10 cannot contradict each other.
-  await admitThreadParticipants({
-    userId,
-    threadId,
-    targetCats,
-    threadStore,
-    socketManager,
-    emitPolicy: 'membership-changed',
-  });
-
-  try {
-    const started = await queueProcessor.processNext(threadId, userId);
-    if (!started.started) {
-      return {
-        messageId: storedMessageId,
-        warning: 'initialMessage queued but did not start automatically',
-      };
-    }
-  } catch (err) {
-    return {
-      messageId: storedMessageId,
-      warning: `initialMessage queued but auto-start failed: ${err instanceof Error ? err.message : String(err)}`,
-    };
-  }
-
-  return { messageId: storedMessageId };
 }
